airflow_orchestration_layer/requirements-airflow.txt
dbt-snowflake==1.11.4
astronomer-cosmos
snowflake-connector-python==4.4.0
psycopg2-binary
asyncpg
requests
polars
azure-keyvault-secrets
azure-identity
python-dotenv
apache-airflow-providers-standard